Addressed CR.
