+++ pyserini/search/_base.py
@@ -1,0 +1,120 @@
+#
+# Pyserini: Python interface to the Anserini IR toolkit built on Lucene
+#
+# Licensed under the Apache License, Version 2.0 (the "License");
+# you may not use this file except in compliance with the License.
+# You may obtain a copy of the License at
+#
+# http://www.apache.org/licenses/LICENSE-2.0
+#
+# Unless required by applicable law or agreed to in writing, software
+# distributed under the License is distributed on an "AS IS" BASIS,
+# WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+# See the License for the specific language governing permissions and
+# limitations under the License.
+#
+
+"""
+This module provides Pyserini's Python search interface to Anserini. The main entry point is the ``SimpleSearcher``
+class, which wraps the Java class with the same name in Anserini.
+"""
+
+import logging
+
+from ..pyclass import JDocument, JTopics, JTopicReader
+
+logger = logging.getLogger(__name__)
+
+
+class Document:
+    """Wrapper class for a Lucene ``Document``.
+
+    Parameters
+    ----------
+    document : JDocument
+        Underlying Lucene ``Document``.
+    """
+
+    def __init__(self, document):
+        if document is None:
+            raise ValueError('Cannot create a Document with None.')
+        self.object = document
+
+    def docid(self: JDocument) -> str:
+        return self.object.getField('id').stringValue()
+
+    def id(self: JDocument) -> str:
+        # Convenient alias for docid()
+        return self.object.getField('id').stringValue()
+
+    def lucene_document(self: JDocument) -> JDocument:
+        return self.object
+
+    def contents(self: JDocument) -> str:
+        return self.object.get('contents')
+
+    def raw(self: JDocument) -> str:
+        return self.object.get('raw')
+
+    def get(self: JDocument, field: str) -> str:
+        return self.object.get(field)
+
+
+def get_topics(collection_name):
+    """
+    Parameters
+    ----------
+    collection_name : str
+        collection_name
+
+    Returns
+    -------
+    result : dictionary
+        Topics as a dictionary
+    """
+    topics = None
+    if collection_name == 'robust04':
+        topics = JTopicReader.getTopicsWithStringIds(JTopics.ROBUST04)
+    elif collection_name == 'robust05':
+        topics = JTopicReader.getTopicsWithStringIds(JTopics.ROBUST05)
+    elif collection_name == 'core17':
+        topics = JTopicReader.getTopicsWithStringIds(JTopics.CORE17)
+    elif collection_name == 'core18':
+        topics = JTopicReader.getTopicsWithStringIds(JTopics.CORE18)
+    elif collection_name == 'car17v1.5_benchmarkY1test':
+        topics = JTopicReader.getTopicsWithStringIds(JTopics.CAR17V15_BENCHMARK_Y1_TEST)
+    elif collection_name == 'car17v2.0_benchmarkY1test':
+        topics = JTopicReader.getTopicsWithStringIds(JTopics.CAR17V20_BENCHMARK_Y1_TEST)
+    elif collection_name == 'msmarco_doc_dev':
+        topics = JTopicReader.getTopicsWithStringIds(JTopics.MSMARCO_DOC_DEV)
+    elif collection_name == 'msmarco_passage_dev_subset':
+        topics = JTopicReader.getTopicsWithStringIds(JTopics.MSMARCO_PASSAGE_DEV_SUBSET)
+    elif collection_name == 'covid_round1':
+        topics = JTopicReader.getTopicsWithStringIds(JTopics.COVID_ROUND1)
+    elif collection_name == 'covid_round1_udel':
+        topics = JTopicReader.getTopicsWithStringIds(JTopics.COVID_ROUND1_UDEL)
+    elif collection_name == 'covid_round2':
+        topics = JTopicReader.getTopicsWithStringIds(JTopics.COVID_ROUND2)
+    elif collection_name == 'covid_round2_udel':
+        topics = JTopicReader.getTopicsWithStringIds(JTopics.COVID_ROUND2_UDEL)
+    elif collection_name == 'covid_round3':
+        topics = JTopicReader.getTopicsWithStringIds(JTopics.COVID_ROUND3)
+    elif collection_name == 'covid_round3_udel':
+        topics = JTopicReader.getTopicsWithStringIds(JTopics.COVID_ROUND3_UDEL)
+    elif collection_name == 'trec2018_bl':
+        topics = JTopicReader.getTopicsWithStringIds(JTopics.TREC2018_BL)
+    elif collection_name == 'trec2019_bl':
+        topics = JTopicReader.getTopicsWithStringIds(JTopics.TREC2019_BL)
+    else:
+        return {}
+    t = {}
+    for topic in topics.keySet().toArray():
+        # Try and parse the keys into integers
+        try:
+            topic_key = int(topic)
+        except ValueError:
+            topic_key = topic
+        t[topic_key] = {}
+        for key in topics.get(topic).keySet().toArray():
+            t[topic_key][key] = topics.get(topic).get(key)
+    return t
